+++ frontend/src/pages/trip-details/index.tsx
@@ -1,0 +1,3 @@
+export function TripDatailsPage() {
+  return <h1>Trip Details</h1>
+}
